修改注册接口的bug

+++ src/main/java/com/mysecret/cn/service/UserInfoService.java
@@ -18,6 +18,6 @@ public interface UserInfoService extends IService<UserInfo> {
      * @param userInfo 用户信息obj
      * @return 新增成功个数
      */
-    Long addUserInfo(UserInfo userInfo);
+    int addUserInfo(UserInfo userInfo);
 
 }
